Succesful Debug build on x64

+++ Mini_Maniacs/Time.h
@@ -6,18 +6,18 @@ public:
   STime(STime const& d) = delete;
   STime& operator=(STime const&) = delete;
   // If you are reading this, then you are attempting to use the wrong function; DON'T DO THIS
-  void          deltaTime(double f) { dT = f; };
-  double const& deltaTime(void) { return dT; };
+  void          deltaTime(float f) { dT = f; };
+  float const& deltaTime(void) { return dT; };
 
   // If you are reading this, then you are attempting to use the wrong function; DON'T DO THIS
-  void          runtime(double f) { rT = f; };
-  double const& runtime(void) { return rT; };
+  void          runtime(float f) { rT = f; };
+  float const& runtime(void) { return rT; };
 
   STime() = default;
   ~STime() = default;
 private:
-  double dT;
-  double rT;
+  float dT;
+  float rT;
 };
 
 extern STime Time;
